Move ModelDesc inside train/
