+++ docs/modules/graph_builder.rst
@@ -2,7 +2,7 @@ tensorpack.graph_builder package
 ================================
 
 These are some useful functions if you need to write your own trainers.
-Note that they may not be well maintained.
+Otherwise you probably don't need to use them.
 
 .. automodule:: tensorpack.graph_builder
     :members:
