+++ tensorpack/graph_builder/__init__.py
@@ -10,6 +10,8 @@ if STATICA_HACK:
     from .distributed import *
     from .utils import *
 
+from .model_desc import InputDesc, ModelDesc, ModelDescBase  # kept for BC # noqa
+
 from pkgutil import iter_modules
 import os
 import os.path
